backend/ingestion-service/requirements.txt
# Backend Ingestion Service Requirements

# HTTP client and async processing
httpx[http2]==0.27.0  # Async HTTP client
requests==2.31.0      # Synchronous HTTP client as fallback

# HTML parsing and content extraction
beautifulsoup4==4.12.2
lxml==4.9.3
trafilatura==1.6.3    # For content extraction

# Text processing and tokenization
tiktoken==0.5.1       # Token counting for chunking
transformers==4.35.0  # Alternative tokenization

# Vector database
qdrant-client==1.7.0

# Embeddings
cohere==4.3.1

# Async utilities
asyncio
aiofiles==23.2.1

# Configuration and environment
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
python-slugify==8.0.1